+++ CHANGELOG.md
@@ -25,9 +25,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Nanook::Account#info` now also returns the `id` of the account.
 - `Nanook::Account#history` now returns `amount` in NANO by default, takes `unit: :raw` argument to return in raw.
 - `Nanook::Account#info` now returns `balance` and `pending` in NANO by default, takes `unit: :raw` argument to return in raw.
-- `Nanook::Block#info` now also returns the `id` of the block.
-- `Nanook::Wallet#seed` alias method of `#id`.
 - `Nanook::Account#exists?` now checks for open block.
+- `Nanook::Block#info` now also returns the `id` of the block.
+- `Nanook::Wallet#create` now returns a `Nanook::Wallet` instance.
+- `Nanook::Wallet#seed` alias method of `#id`.
+- `Nanook::WalletAccount#create` now returns `Nanook::WalletAccount` instances.
 - Changed documentation generating tool from rdoc to yard.
 
 ### Fixed